auth/requirements.txt
flask==2.3.3
psycopg2-binary==2.9.7
pyjwt==2.8.0
werkzeug==2.3.7